--- Builds/.dockerignore
@@ -1,12 +1,0 @@
-# Ignore everything
-*
-
-# Allow files and directories
-!/scripts
-
-# Ignore unnecessary files inside allowed directories
-# This should go after the allowed directories
-**/*~
-**/*.log
-**/.DS_Store
-**/Thumbs.db
